feat(credential): gate OAuth2 HTTP behind oauth2-http feature
Extract authorization URL construction to oauth2/authorize_url.rs (no reqwest).
Make reqwest optional; enable it by default via the oauth2-http feature so
existing consumers keep full OAuth2 behavior.

When oauth2-http is disabled, token exchange, device polling, and refresh
return CredentialError::Provider with a message pointing at the feature.

CI: cargo check -p nebula-credential --no-default-features.

ADR-0031 incremental split toward relocating HTTP from the contract crate.

Made-with: Cursor
