+++ crates/credential/Cargo.toml
@@ -52,8 +52,12 @@ nebula-resilience = { path = "../resilience" }
 nebula-error = { workspace = true }
 thiserror = { workspace = true }
 
-# HTTP client (for flows)
-reqwest = { workspace = true, default-features = false, features = ["json", "rustls", "form"] }
+# HTTP client (OAuth2 token exchange / device poll / refresh) — optional; see feature `oauth2-http`.
+reqwest = { workspace = true, default-features = false, features = [
+  "json",
+  "rustls",
+  "form",
+], optional = true }
 url = { workspace = true }
 
 # Serialization
@@ -82,7 +86,10 @@ tracing-subscriber = { workspace = true, features = ["fmt"] }
 semver = { workspace = true }
 
 [features]
-default = []
+default = ["oauth2-http"]
+# Token exchange, device-code polling, and refresh over HTTP (`reqwest`). When disabled, the crate
+# has no HTTP client dependency; flows that need network I/O return `CredentialError::Provider`.
+oauth2-http = ["dep:reqwest"]
 rotation = []
 # Expose test-only helpers (e.g. `StaticKeyProvider`) to integration tests and
 # to other crates' dev-dependency paths. Must never be enabled in a production
